+++ src/lib/components/GraphInterface.svelte
@@ -3,8 +3,8 @@
 
 
 	// Lib imports
-    import { Graph } from '$scripts/graph/entities'
-	import { GraphSVG, GraphType } from '$scripts/graph/graphSVG'
+    import { Graph } from '$scripts/entities'
+	import { GraphSVG, GraphType } from '$scripts/d3'
 	
 	// Components
 	import Dropdown from './Dropdown.svelte';
@@ -41,8 +41,8 @@
 
         <button
 			class="tab last"
-            class:active={graphSVG.type === GraphType.lecture}
-            on:click={() => graphSVG.type = GraphType.lecture}
+            class:active={graphSVG.type === GraphType.lectures}
+            on:click={() => graphSVG.type = GraphType.lectures}
         > Lectures </button>
 
 		<Dropdown label="Lecture" placeholder="Choose a Lecture" bind:value={graphSVG.lecture} options={graph.lectureOptions} />
