+++ src/lib/scripts/d3/FieldSVG.ts
@@ -3,10 +3,10 @@
 import * as d3 from 'd3'
 
 // Internal imports
-import { Field, Relation } from './entities'
-import { RelationSVG } from './relationSVG'
-import * as settings from './settings'
-import { styles } from './settings'
+import { RelationSVG } from '../d3'
+import { Field, Relation } from '../entities'
+import * as settings from '../settings'
+import { styles } from '../settings'
 
 // Exports
 export { FieldSVG }
