+++ src/lib/scripts/d3/index.ts
@@ -1,0 +1,4 @@
+
+export { GraphSVG, GraphType } from './GraphSVG'
+export { FieldSVG } from './FieldSVG'
+export { RelationSVG } from './RelationSVG'
