+++ src/lib/scripts/entities/index.ts
@@ -1,0 +1,6 @@
+
+export { Course } from './Course'
+export { Graph } from './Graph'
+export { Field, Domain, Subject } from './Fields'
+export { Relation } from './Relations'
+export { Lecture } from './Lectures'
