+++ src/lib/scripts/settings.ts
